# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "firstName": "Juan",
                    "lastName": "Pérez",
                    "rut": "12345678-9",
                    "email": "juan@climas.com",
                    "phone": "+56912345678",
                    "roles": [
                      "ADMIN"
                    ],
                    "isActive": true
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Se requiere rol ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Usuario no encontrado",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener un usuario por ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "firstName": "Juan",
                    "lastName": "Pérez",
                    "rut": "12345678-9",
                    "email": "juan@climas.com",
                    "phone": "+56912345678",
                    "roles": [
                      "ADMIN"
                    ],
                    "isActive": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Usuario no encontrado",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Se requiere rol ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar un usuario",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Usuario eliminado exitosamente"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Se requiere rol ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un usuario",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/consultation": {
      "post": {
        "description": "Programa una nueva cita médica en el sistema.\n\n### 🛑 Importante: Flujo a seguir\nPara evitar errores al crear una consulta, asegúrate de:\n1. **Obtener Doctor**: Consultar `GET /doctor` para obtener el **ID real del doctor** y enviarlo en `doctorId`. (No usar el ID de `User`).\n2. **Obtener Paciente**: Consultar `GET /patient` para obtener el **ID real del paciente** y enviarlo en `patientId`. (No usar el ID de `User`).\n3. **Servicios (Opcional)**: Consultar `GET /service` para extraer los IDs de los servicios y enviarlos en `serviceIds` como un arreglo de strings.\n\n**Roles permitidos**: ADMIN y DOCTOR.",
        "operationId": "ConsultationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsultationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Consulta creada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "cuid67890",
                    "dateTime": "2026-06-10T15:30:00.000Z",
                    "reason": "Chequeo general",
                    "status": "SCHEDULED",
                    "doctorId": "cuid-doc-123",
                    "patientId": "cuid-pat-456",
                    "createdAt": "2026-06-03T10:00:00.000Z",
                    "updatedAt": "2026-06-03T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos (Bad Request). Ej: Formato de fecha incorrecto o serviceIds no es un array.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": [
                      "dateTime must be a Date instance"
                    ],
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Se requiere Token JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Prohibido. Se requiere rol ADMIN o DOCTOR.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor. Suele ocurrir por violación de Foreign Key (ej. enviar el ID de un User en lugar del de un Doctor o Paciente).",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 500,
                    "message": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear una nueva consulta",
        "tags": [
          "Consultation"
        ]
      },
      "get": {
        "description": "Obtiene el registro histórico completo de todas las consultas médicas del sistema.\n    \n**Incluye relaciones:** \n- Datos básicos del doctor.\n- Datos básicos del paciente.\n- Servicios brindados.\n- Registro Médico (clinical notes, prescription, diagnosis) asociado.\n\n**Roles permitidos**: ADMIN.",
        "operationId": "ConsultationController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de consultas obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "cuid67890",
                      "dateTime": "2026-06-10T15:30:00.000Z",
                      "reason": "Chequeo general",
                      "status": "SCHEDULED",
                      "doctorId": "cuid-doc-123",
                      "patientId": "cuid-pat-456",
                      "createdAt": "2026-06-03T10:00:00.000Z",
                      "updatedAt": "2026-06-03T10:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "No hay consultas registradas.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "No hay consultas registradas"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Se requiere Token JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Prohibido. Se requiere rol ADMIN.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener todas las consultas",
        "tags": [
          "Consultation"
        ]
      }
    },
    "/consultation/user/{userId}": {
      "get": {
        "description": "Retorna todas las citas asociadas a un **usuario** en específico. \n    \nEste endpoint busca inteligentemente en la base de datos para saber si el usuario participó en la cita médica, **ya sea como Doctor o como Paciente**.\n\n**Roles permitidos**: ADMIN, DOCTOR, PATIENT.",
        "operationId": "ConsultationController_findByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consultas del usuario obtenidas exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "cuid67890",
                      "dateTime": "2026-06-10T15:30:00.000Z",
                      "reason": "Chequeo general",
                      "status": "SCHEDULED",
                      "doctorId": "cuid-doc-123",
                      "patientId": "cuid-pat-456",
                      "createdAt": "2026-06-03T10:00:00.000Z",
                      "updatedAt": "2026-06-03T10:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "No hay consultas para este usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "No hay consultas para este usuario"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. Se requiere Token JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener consultas por ID de usuario (User ID)",
        "tags": [
          "Consultation"
        ]
      }
    },
...
  },
  "info": {
    "title": "API CLIMAS",
    "description": "Documentación API CLIMAS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "rut": {
            "type": "string",
            "example": "12345678-9"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "johndoe@climas.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "example": "MiSuperPassword123"
          },
          "phone": {
            "type": "string",
            "example": "+56912345678"
          },
          "roles": {
            "type": "string",
            "items": {
              "type": "string",
              "enum": [
                "PATIENT",
                "ADMIN",
                "DOCTOR",
                "WORKER"
              ]
            },
            "example": [
              "PATIENT"
            ],
            "enum": [
              "ADMIN",
              "DOCTOR",
              "WORKER",
              "PATIENT"
            ]
          }
        }
      },
...
      "CreateConsultationDto": {
        "type": "object",
        "properties": {
          "dateTime": {
            "type": "object",
            "description": "Fecha y hora de la consulta médica (formato ISO 8601).",
            "example": "2026-06-02T10:00:00.000Z"
          },
          "reason": {
            "type": "string",
            "description": "Motivo principal de la consulta. Ayuda al doctor a tener contexto previo a la cita.",
            "example": "Dolor de espalda agudo y persistente"
          },
          "status": {
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELED",
              "NO_SHOW"
            ],
            "type": "string",
            "description": "Estado actual de la consulta. Por defecto, al crear una cita nueva el estado es `SCHEDULED`.",
            "default": "SCHEDULED"
          },
          "doctorId": {
            "type": "string",
            "description": "El ID único del doctor. **¡Atención!** Este debe ser el ID de la tabla `Doctor`, **NO** el ID del `User` asociado a ese doctor.",
            "example": "cuid-doctor-123"
          },
          "patientId": {
            "type": "string",
            "description": "El ID único del paciente. **¡Atención!** Este debe ser el ID de la tabla `Patient`, **NO** el ID del `User` asociado a ese paciente.",
            "example": "cuid-patient-123"
          },
          "serviceIds": {
            "type": "array",
            "items": {
              "required": false,
              "description": "Arreglo de IDs de los servicios a realizar durante la consulta. Debe ser un arreglo simple de strings.",
              "example": [
                "cuid-service-1"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "dateTime",
          "doctorId",
          "patientId",
          "serviceIds"
        ]
      },
...
    }
  }
}